+++ src/services/orders.service.ts
@@ -1,60 +1,76 @@
-import { HttpException } from '@exceptions/HttpException';
-import ordersModel from '@models/orders.model';
-import { isEmpty } from '@utils/util';
-import { OrderDto } from '@dtos/orders.dto';
+import { HttpException } from "@exceptions/HttpException";
+import ordersModel from "@models/orders.model";
+import { isEmpty } from "@utils/util";
+import { OrderDto } from "@dtos/orders.dto";
+import { IOrderQuery, IOrderQueryOptions } from "@interfaces/orders.interface";
 
 class UserService {
-  public orders = ordersModel;
+    public orders = ordersModel;
 
-  public async findAllOrders(): Promise<OrderDto[]> {
-    const orders: OrderDto[] = await this.orders
-      .find()
-      .populate('order_list.format')
-      .populate('order_list.film')
-      .populate('order_list.responsible')
-      .populate('last_edit')
-      .populate('comments.author');
-    return orders;
-  }
+    public async findAllOrders(
+        query?: IOrderQuery,
+        options?: IOrderQueryOptions
+    ): Promise<OrderDto[]> {
+        const orders: OrderDto[] = await this.orders
+            .find({ ...query })
+			.skip(options.skip)
+			.limit(options.limit)
+            .populate("order_list.format")
+            .populate("order_list.film")
+            .populate("order_list.responsible")
+            .populate("last_edit")
+            .populate("comments.author");
+        return orders;
+    }
 
-  public async findOrderById(orderId: string): Promise<OrderDto> {
-    if (isEmpty(orderId)) throw new HttpException(400, 'Bad request');
+    public async findOrderById(orderId: string): Promise<OrderDto> {
+        if (isEmpty(orderId)) throw new HttpException(400, "Bad request");
 
-    const findOrder: OrderDto = await this.orders
-      .findOne({ _id: orderId })
-      .populate('order_list.format')
-      .populate('order_list.film')
-      .populate('order_list.responsible')
-      .populate('last_edit')
-      .populate('comments.author');
-    if (!findOrder) throw new HttpException(404, 'Order not found');
+        const findOrder: OrderDto = await this.orders
+            .findOne({ _id: orderId })
+            .populate("order_list.format")
+            .populate("order_list.film")
+            .populate("order_list.responsible")
+            .populate("last_edit")
+            .populate("comments.author");
+        if (!findOrder) throw new HttpException(404, "Order not found");
 
-    return findOrder;
-  }
+        return findOrder;
+    }
 
-  public async createOrder(orderData: OrderDto): Promise<OrderDto> {
-    if (isEmpty(orderData)) throw new HttpException(400, 'Bad request');
+    public async createOrder(orderData: OrderDto): Promise<OrderDto> {
+        if (isEmpty(orderData)) throw new HttpException(400, "Bad request");
 
-    const createOrderData: OrderDto = await this.orders.create({ ...orderData });
+        const createOrderData: OrderDto = await this.orders.create({
+            ...orderData,
+        });
 
-    return createOrderData;
-  }
+        return createOrderData;
+    }
 
-  public async updateOrder(orderId: string, orderData: OrderDto): Promise<OrderDto> {
-    if (isEmpty(orderData)) throw new HttpException(400, 'Bad request');
+    public async updateOrder(
+        orderId: string,
+        orderData: OrderDto
+    ): Promise<OrderDto> {
+        if (isEmpty(orderData)) throw new HttpException(400, "Bad request");
 
-    const updateOrderById: OrderDto = await this.orders.findByIdAndUpdate(orderId, { ...orderData });
-    if (!updateOrderById) throw new HttpException(404, 'Order not found');
+        const updateOrderById: OrderDto = await this.orders.findByIdAndUpdate(
+            orderId,
+            { ...orderData }
+        );
+        if (!updateOrderById) throw new HttpException(404, "Order not found");
 
-    return updateOrderById;
-  }
+        return updateOrderById;
+    }
 
-  public async deleteOrder(orderId: string): Promise<OrderDto> {
-    const deleteUserById: OrderDto = await this.orders.findByIdAndDelete(orderId);
-    if (!deleteUserById) throw new HttpException(404, 'Order not found');
+    public async deleteOrder(orderId: string): Promise<OrderDto> {
+        const deleteUserById: OrderDto = await this.orders.findByIdAndDelete(
+            orderId
+        );
+        if (!deleteUserById) throw new HttpException(404, "Order not found");
 
-    return deleteUserById;
-  }
+        return deleteUserById;
+    }
 }
 
 export default UserService;